+++ lmdeploy/pytorch/devices/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .device_manager import DefaultContext, DeviceContext, get_device_manager
+
+__all__ = ['DeviceContext', 'DefaultContext', 'get_device_manager']
