+++ lmdeploy/pytorch/engine/engine.py
@@ -14,6 +14,7 @@ from lmdeploy.utils import get_logger, get_model, logging_timer
 from ..adapter.adapter import AdapterManager, SchedulerAdapter
 from ..check_env import check_adapters, check_env, check_model
 from ..config import CacheConfig, SchedulerConfig
+from ..devices import DeviceContext, get_device_manager
 from ..messages import MessageStatus, SchedulerSequence
 from ..paging import Scheduler
 from .logits_process import FusedLogitsProcessor, SamplingInputs
@@ -94,7 +95,7 @@ class Engine:
                  model_path: str,
                  engine_config: PytorchEngineConfig = None,
                  trust_remote_code: bool = True) -> None:
-        check_env()
+        check_env(engine_config.device_type)
         check_model(model_path, trust_remote_code)
         if engine_config is None:
             engine_config = PytorchEngineConfig()
@@ -107,6 +108,9 @@ class Engine:
 
         self.tp = tp
         self.model_name = model_name
+
+        self.device_context = DeviceContext(
+            device_type=engine_config.device_type)
 
         scheduler_config = SchedulerConfig(
             max_batches=engine_config.max_batch_size,
@@ -130,12 +134,13 @@ class Engine:
                                    engine_config.revision)
         self.model_path = model_path
 
-        self.model_agent = AutoModelAgent.from_pretrained(
-            model_path,
-            cache_config=cache_config,
-            trust_remote_code=trust_remote_code,
-            adapters=adapters,
-            tp=tp)
+        with get_device_manager().context(self.device_context):
+            self.model_agent = AutoModelAgent.from_pretrained(
+                model_path,
+                cache_config=cache_config,
+                trust_remote_code=trust_remote_code,
+                adapters=adapters,
+                tp=tp)
 
         cache_config = self.model_agent.cache_config
         self.adapter_manager = self._build_adapter_manager(adapters)
@@ -681,24 +686,23 @@ class Engine:
                 self._running = running
                 self._inputs = inputs
 
-                with torch.cuda.stream(self.stream):
-                    await self._async_step_background(
-                        inputs=inputs,
-                        swap_in_map=schedule_output.swap_in_map,
-                        swap_out_map=schedule_output.swap_out_map,
-                        history_ids=history_ids,
-                        sampling_inputs=sampling_inputs,
-                        num_appendable_ids=num_appendable_ids,
-                        num_ignore_eos=num_ignore_eos,
-                        output_que=out_que,
-                    )
+                await self._async_step_background(
+                    inputs=inputs,
+                    swap_in_map=schedule_output.swap_in_map,
+                    swap_out_map=schedule_output.swap_out_map,
+                    history_ids=history_ids,
+                    sampling_inputs=sampling_inputs,
+                    num_appendable_ids=num_appendable_ids,
+                    num_ignore_eos=num_ignore_eos,
+                    output_que=out_que,
+                )
             except Exception as e:
                 out_que.put_nowait((True, e))
             finally:
                 in_que.task_done()
 
     @torch.inference_mode()
-    async def async_loop(self):
+    async def _async_loop(self):
         """Main loop of the engine.
 
         Each engine instance would communicate with the engine by queue.
@@ -760,6 +764,12 @@ class Engine:
             # decoding
             if self.scheduler.has_running():
                 await __step(False)
+
+    async def async_loop(self):
+        device_manager = get_device_manager()
+        with device_manager.context(self.device_context), torch.cuda.stream(
+                self.stream):
+            await self._async_loop()
 
     def create_instance(self, cuda_stream_id=0):
         """Create a pytorch engine instance.
