+++ lmdeploy/pytorch/engine/model_agent.py
@@ -17,6 +17,7 @@ from ..adapter.adapter import (AdapterWeightMap, SchedulerAdapter,
                                get_indexed_lora_linears, get_loralinear_info,
                                update_lora_linears)
 from ..config import CacheConfig, ModelConfig
+from ..devices import DeviceContext, get_device_manager
 from ..models import patch
 from ..utils import get_gpu_memory
 from .cache_engine import CacheEngine
@@ -908,6 +909,7 @@ def _tp_model_loop(
 def _start_tp_process(proc_id: int,
                       world_size: int,
                       func: Callable,
+                      device_context: DeviceContext,
                       args: List = None,
                       kwargs: Dict = None):
     """Start the tensor parallel process.
@@ -925,7 +927,9 @@ def _start_tp_process(proc_id: int,
                                 rank=rank,
                                 world_size=world_size,
                                 timeout=timedelta(days=35600))
-        with torch.cuda.device(rank), torch.inference_mode():
+        torch.cuda.set_device(rank)
+        with get_device_manager().context(
+                device_context), torch.inference_mode():
             args = args or tuple()
             kwargs = kwargs or dict()
             func(rank, *args, **kwargs)
@@ -1033,11 +1037,13 @@ class TPModelAgent(AutoModelAgent):
         port = os.environ['MASTER_PORT']
         logger.info(f'MASTER_ADDR={addr}, MASTER_PORT={port}')
 
+        device_context = get_device_manager().current_context()
         self.mp_context = mp.spawn(
             _start_tp_process,
             args=(
                 world_size,
                 _tp_model_loop,
+                device_context,
                 (model_path, ),
                 dict(model_config=model_config,
                      cache_config=cache_config,
