+++ lmdeploy/pytorch/models/functional.py
@@ -1,13 +1,10 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import math
-from typing import Any, Callable, Optional, Sequence, Tuple
+from typing import Any
 
-import numpy as np
 import torch
-# import torch.nn.functional as F
-from torch import Tensor
 
-from ..kernels import apply_rotary_pos_emb, fill_kv_cache, rerope_attention_fwd
+from ..kernels import apply_rotary_pos_emb
 
 __all__ = ['apply_rotary_pos_emb']
 
@@ -160,204 +157,3 @@ def sync_qparam_to_context(context: Any, layer_id: str, qparams: dict):
         context.set_output(layer_id, last_qparam)
     else:
         context.set_output(layer_id, qparams)
-
-
-@torch.no_grad()
-def attention_forward_with_rerope(
-        hidden_states: Tensor,
-        history_lengths: Sequence,
-        block_offsets: Tensor,
-        num_heads: int,
-        num_kv_heads: int,
-        head_dim: int,
-        position_ids: torch.LongTensor,
-        past_key_value: Tuple[Tensor],
-        attention_mask: Tensor,
-        context: Any = None,
-        q_proj: Optional[Callable] = None,
-        k_proj: Optional[Callable] = None,
-        v_proj: Optional[Callable] = None,
-        qkv_proj: Optional[Callable] = None,
-        o_proj: Optional[Callable] = None,
-        rotary_emb_context_fn: Optional[Callable] = None,
-        rotary_emb_generate_fn: Optional[Callable] = None,
-        bias_type: str = 'default',
-        training_length=4096,
-        window=512,
-        layer_id: str = None) -> Tensor:
-    """Attention module forward with ReRoPE.
-
-    Args:
-        hidden_states (Tensor): Input of attention layer.
-        history_lengths (Sequence): Cache lengths of each data in batch.
-        block_offsets (Tensor): Block table of the key/value caches,
-            used by paged attention.
-        num_heads (int): numbers of query heads.
-        num_kv_heads (int): numbers of key/value heads.
-        head_dim (int): Feature dimension of heads.
-        position_ids (LongTensor): position ids of the input.
-        past_key_value (Tuple[Tensor]): key value cache.
-        q_proj (Callable): query project module/function.
-        k_proj (Callable): key project module/function.
-        v_proj (Callable): value project module/function.
-        qkv_proj (Callable): query/key/value project module/function.
-        o_proj (Callable): output project module/function.
-        rotary_emb_context_fn (Callable): rotary embedding context callback.
-        rotary_emb_generate_fn (Callable): rotary embedding generate callback.
-        bias_type (str): type of attention bias. support ['default'].
-        training_length (int): model sequence length during trainning.
-        window (int): ReRoPE window size, default value is 512.
-    """
-    hidden_size = -1
-    if qkv_proj is not None:
-        assert q_proj is None
-        assert k_proj is None
-        assert v_proj is None
-        query_states, key_states, value_states = qkv_proj(hidden_states)
-    else:
-        assert qkv_proj is None
-        assert q_proj is not None
-        assert k_proj is not None
-        assert v_proj is not None
-        query_states = q_proj(hidden_states)
-        key_states = k_proj(hidden_states)
-        value_states = v_proj(hidden_states)
-
-    hidden_size = num_heads * head_dim
-    query_states = query_states.view(-1, num_heads, head_dim)
-    key_states = key_states.view(-1, num_kv_heads, head_dim)
-    value_states = value_states.view(-1, num_kv_heads, head_dim)
-    query_states *= ((position_ids.flatten() + 1)[:, None, None].log() /
-                     np.log(training_length)).clip(1).to(query_states.dtype)
-
-    kv_seq_length = (position_ids[..., -1] + 1).item()
-
-    q_seq_length = getattr(context, 'q_seq_length', None)
-    if q_seq_length is None:
-        q_seq_length = kv_seq_length - kv_seq_length.new_tensor(
-            history_lengths)
-
-    q_start_loc = getattr(context, 'q_start_loc', None)
-    if q_start_loc is None:
-        q_start_loc = q_seq_length.cumsum(0)
-        q_start_loc = torch.cat([q_start_loc.new_zeros(1), q_start_loc[:-1]])
-
-    if past_key_value[0].dtype != hidden_states.dtype:
-        # dynamic quantize hidden_states to kv_cache and save
-        quant = True
-        qkey, qvalue, qparams = quant_kv(key_states, value_states,
-                                         past_key_value[0].dtype)
-        sync_qparam_to_context(context=context,
-                               layer_id=layer_id,
-                               qparams=qparams)
-
-        fill_kv_cache(qkey,
-                      qvalue,
-                      past_key_value[0],
-                      past_key_value[1],
-                      q_start_loc,
-                      q_seq_length,
-                      block_offsets=block_offsets,
-                      history_lengths=history_lengths,
-                      context=context)
-
-    else:
-        fill_kv_cache(key_states,
-                      value_states,
-                      past_key_value[0],
-                      past_key_value[1],
-                      q_start_loc,
-                      q_seq_length,
-                      block_offsets=block_offsets,
-                      history_lengths=history_lengths,
-                      context=context)
-
-    bsz, q_len, _ = hidden_states.size()
-    if bias_type.lower() == 'default':
-
-        if q_len == 1:
-
-            key_states = past_key_value[0][block_offsets].view(
-                -1, num_heads, head_dim)[0:history_lengths[-1] + 1]
-            value_states = past_key_value[1][block_offsets].view(
-                -1, num_heads, head_dim)[0:history_lengths[-1] + 1]
-
-            if quant:
-                # dequant int8 tensor to hidden_states.dtype
-                key_states, value_states = dequant_kv(
-                    context=context,
-                    layer_id=layer_id,
-                    key_int8=key_states,
-                    value_int8=value_states,
-                    out_type=hidden_states.dtype)
-
-            full_position_ids = torch.arange(
-                position_ids.item() + 1,
-                device=position_ids.device).unsqueeze(0)
-
-            key_states, value_states = rotary_emb_generate_fn(
-                key_states, value_states, full_position_ids, window)
-            attn_weights = torch.matmul(query_states.transpose(
-                0, 1), key_states.permute(1, 2, 0)) / math.sqrt(head_dim)
-
-            if attention_mask is not None:
-                attn_weights = attn_weights + attention_mask
-
-            # upcast attention to fp32
-            attn_weights = torch.nn.functional.softmax(attn_weights,
-                                                       dim=-1,
-                                                       dtype=torch.float32).to(
-                                                           query_states.dtype)
-            attn_output = torch.matmul(attn_weights,
-                                       value_states.transpose(0, 1))
-
-        else:
-            query_states1, query_states2, key_states1, key_states2, value_states = rotary_emb_context_fn(  # noqa: E501
-                query_states, key_states, value_states, position_ids, window)
-
-            sm_scale = 1.0 / math.sqrt(head_dim)
-
-            PADDING_UNIT = past_key_value[0].shape[1]
-            assert PADDING_UNIT in {16, 32, 64, 128, 256}
-            # padding_len = -query_states1.shape[2] % PADDING_UNIT
-            # query_states1 = F.pad(query_states1,
-            #                       (0, 0, 0, padding_len)).contiguous()
-            # query_states2 = F.pad(query_states2,
-            #                       (0, 0, 0, padding_len)).contiguous()
-            # key_states1 = F.pad(key_states1,
-            #                     (0, 0, 0, padding_len)).contiguous()
-            # key_states2 = F.pad(key_states2,
-            #                     (0, 0, 0, padding_len)).contiguous()
-            # value_states = F.pad(value_states,
-            #                      (0, 0, 0, padding_len)).contiguous()
-
-            query_states1 = query_states1.contiguous()
-            query_states2 = query_states2.contiguous()
-            key_states1 = key_states1.contiguous()
-            key_states2 = key_states2.contiguous()
-            value_states = value_states.contiguous()
-            attn_output = rerope_attention_fwd(query_states1,
-                                               query_states2,
-                                               key_states1,
-                                               key_states2,
-                                               value_states,
-                                               True,
-                                               sm_scale,
-                                               window,
-                                               BLOCK_M=PADDING_UNIT).squeeze(0)
-
-            # attn_output = attn_output[:, 0:q_len]
-
-        if attn_output.size() != (num_heads, q_len, head_dim):
-            raise ValueError(
-                f'`attn_output` should be of size {(bsz, num_heads, q_len, head_dim)}, but is'  # noqa: E501
-                f' {attn_output.size()}')
-
-        attn_output = attn_output.transpose(0, 1).reshape(
-            bsz, q_len, hidden_size).contiguous()
-    else:
-        raise ValueError(f'Unknown bias type: {bias_type}')
-
-    if o_proj is not None:
-        attn_output = o_proj(attn_output)
-    return attn_output
